backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
alembic==1.14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.7.1
python-multipart==0.0.20
email-validator==2.2.0
pytest==8.3.4
httpx==0.28.1
